requirements.txt
typer>=0.9.0 
aiohttp>=3.8.0 
pyyaml>=6.0.0 
pytest>=7.0.0 
pytest-asyncio>=0.21.0
rich>=13.0.0
typer==0.12.5
requests==2.32.3
pyyaml==6.0.2
rich==13.9.2
beautifulsoup4==4.12.3
aiohttp==3.10.5
watchdog==5.0.3